Add a PUT api endpoint to postRoutes.js
In this commit, I added a PUT endpoint which gets triggered for routes which matches the router.put(api/posts/:id) url. A handler callback function is invoked.
The function does the following
1. retrieves the id of the post from req.params
2. gets the title and contents of the post from req.body
3. checks if title or the content of the post which will be used for updating existing post is empty (if true, throws an err)
creates a new post obj and overrides existing post with it
return 1 if update is successful or 0 otherwise