Add a GET/:id api endpoint to postRoutes.js
In this commit, I added a GET endpoint which gets triggered for get requests which matches the url router.get(api/posts/:id) url. A handler callback function is invoked.
The function does the following
1. retrieves the id of the post from req.params
2. checks to see if the post with the specified id exists.
If the post exists, the post object is returned otherwise we
throw an error saying the post with that id does not exist