Add a GET/:id api endpoint to postRoutes.js
In this commit, I added a GET endpoint which gets triggered for get requests which matches the url router.get(api/posts/:id/comments) url. This request return a list of comments for the post with the specified id. When this get route is made triggered, we invoked a function which does the following

1. retrieves the id of the post from req.params and invokes the findById() method passing the id of the post
2. checks to see if the post with the specified id exists, we then proceed to fetch all the comments associated with that post